Implement slide-step error showing
Also, refactor the slide-step form/error showing so that upon
continuing, we remove the address validation content and show
the native shipping form (based on a timeout of two seconds).

That way, after the step is completed, a user can click back on the
shipping information tab and enter their info normally.