Refactor AjaxController
Things are more dynamic now, since each request will be so similar.
There is an _abstractAction that each actual methodAction calls on, and
results will differ slightly depending on $request->getAction(). I tried
to make a __call(), with an $_allowedActions array, but the router
requires the methodAction to exist on the class.